lot: merge LOT modules into TRX
This merges the two games' LOT modules into TRX, retaining the
differences for slot initialisation based on enemy type. This will be
further refactored later.
