+++ src/libtrx/game/pathing/lot.c
@@ -1,13 +1,11 @@
-#include "game/lot.h"
+#include "game/pathing/lot.h"
 
-#include "game/box.h"
-#include "game/shell.h"
-#include "global/vars.h"
-
-#include <libtrx/debug.h>
-#include <libtrx/game/camera.h>
-#include <libtrx/game/game_buf.h>
-#include <libtrx/utils.h>
+#include "debug.h"
+#include "game/camera.h"
+#include "game/game_buf.h"
+#include "game/pathing.h"
+#include "game/rooms.h"
+#include "utils.h"
 
 static int32_t m_SlotsUsed = 0;
 static CREATURE *m_BaddieSlots = nullptr;
@@ -16,12 +14,14 @@ void LOT_InitialiseArray(void)
 {
     m_BaddieSlots =
         GameBuf_Alloc(LOT_SLOT_COUNT * sizeof(CREATURE), GBUF_CREATURE_DATA);
-    for (int i = 0; i < LOT_SLOT_COUNT; i++) {
-        CREATURE *creature = &m_BaddieSlots[i];
+
+    for (int32_t i = 0; i < LOT_SLOT_COUNT; i++) {
+        CREATURE *const creature = &m_BaddieSlots[i];
         creature->item_num = NO_ITEM;
         creature->lot.node =
-            GameBuf_Alloc(sizeof(BOX_NODE) * Box_GetCount(), GBUF_CREATURE_LOT);
+            GameBuf_Alloc(Box_GetCount() * sizeof(BOX_NODE), GBUF_CREATURE_LOT);
     }
+
     m_SlotsUsed = 0;
 }
 
@@ -30,12 +30,13 @@ CREATURE *LOT_GetBaddieSlot(const int32_t i)
     return &m_BaddieSlots[i];
 }
 
-void LOT_DisableBaddieAI(int16_t item_num)
+void LOT_DisableBaddieAI(const int16_t item_num)
 {
     ITEM *const item = Item_Get(item_num);
-    CREATURE *creature = item->data;
+    CREATURE *const creature = (CREATURE *)item->data;
     item->data = nullptr;
-    if (creature) {
+
+    if (creature != nullptr) {
         creature->item_num = NO_ITEM;
         m_SlotsUsed--;
     }
@@ -49,8 +50,7 @@ bool LOT_EnableBaddieAI(const int16_t item_num, const bool always)
 
     if (m_SlotsUsed < LOT_SLOT_COUNT) {
         for (int32_t slot = 0; slot < LOT_SLOT_COUNT; slot++) {
-            CREATURE *creature = &m_BaddieSlots[slot];
-            if (creature->item_num == NO_ITEM) {
+            if (m_BaddieSlots[slot].item_num == NO_ITEM) {
                 LOT_InitialiseSlot(item_num, slot);
                 return true;
             }
@@ -61,20 +61,20 @@ bool LOT_EnableBaddieAI(const int16_t item_num, const bool always)
     int32_t worst_dist = 0;
     if (!always) {
         const ITEM *const item = Item_Get(item_num);
-        int32_t x = (item->pos.x - g_Camera.pos.x) >> 8;
-        int32_t y = (item->pos.y - g_Camera.pos.y) >> 8;
-        int32_t z = (item->pos.z - g_Camera.pos.z) >> 8;
-        worst_dist = SQUARE(x) + SQUARE(y) + SQUARE(z);
+        const int32_t dx = (item->pos.x - g_Camera.pos.pos.x) >> 8;
+        const int32_t dy = (item->pos.y - g_Camera.pos.pos.y) >> 8;
+        const int32_t dz = (item->pos.z - g_Camera.pos.pos.z) >> 8;
+        worst_dist = SQUARE(dx) + SQUARE(dy) + SQUARE(dz);
     }
 
     int32_t worst_slot = -1;
     for (int32_t slot = 0; slot < LOT_SLOT_COUNT; slot++) {
-        CREATURE *creature = &m_BaddieSlots[slot];
-        const ITEM *const item = Item_Get(creature->item_num);
-        int32_t x = (item->pos.x - g_Camera.pos.x) >> 8;
-        int32_t y = (item->pos.y - g_Camera.pos.y) >> 8;
-        int32_t z = (item->pos.z - g_Camera.pos.z) >> 8;
-        int32_t dist = SQUARE(x) + SQUARE(y) + SQUARE(z);
+        const int32_t item_num = m_BaddieSlots[slot].item_num;
+        const ITEM *const item = Item_Get(item_num);
+        const int32_t dx = (item->pos.x - g_Camera.pos.pos.x) >> 8;
+        const int32_t dy = (item->pos.y - g_Camera.pos.pos.y) >> 8;
+        const int32_t dz = (item->pos.z - g_Camera.pos.pos.z) >> 8;
+        const int32_t dist = SQUARE(dx) + SQUARE(dy) + SQUARE(dz);
         if (dist > worst_dist) {
             worst_dist = dist;
             worst_slot = slot;
@@ -85,30 +85,37 @@ bool LOT_EnableBaddieAI(const int16_t item_num, const bool always)
         return false;
     }
 
-    Item_Get(m_BaddieSlots[worst_slot].item_num)->status = IS_INVISIBLE;
-    LOT_DisableBaddieAI(m_BaddieSlots[worst_slot].item_num);
+    const CREATURE *const creature = &m_BaddieSlots[worst_slot];
+    Item_Get(creature->item_num)->status = IS_INVISIBLE;
+    LOT_DisableBaddieAI(creature->item_num);
     LOT_InitialiseSlot(item_num, worst_slot);
     return true;
 }
 
-void LOT_InitialiseSlot(int16_t item_num, int32_t slot)
+void LOT_InitialiseSlot(const int16_t item_num, const int32_t slot)
 {
-    CREATURE *creature = &m_BaddieSlots[slot];
+    CREATURE *const creature = &m_BaddieSlots[slot];
     ITEM *const item = Item_Get(item_num);
     item->data = creature;
+
     creature->item_num = item_num;
     creature->mood = MOOD_BORED;
-    creature->head_rotation = 0;
     creature->neck_rotation = 0;
+    creature->head_rotation = 0;
     creature->maximum_turn = DEG_1;
     creature->flags = 0;
-
+    creature->enemy = nullptr;
     creature->lot.step = STEP_L;
+#if TR_VERSION == 1
     creature->lot.drop = -STEP_L;
+#else
+    creature->lot.drop = -STEP_L * 2;
+#endif
     creature->lot.block_mask = BOX_BLOCKED;
     creature->lot.fly = 0;
 
     switch (item->object_id) {
+#if TR_VERSION == 1
     case O_BAT:
     case O_ALLIGATOR:
     case O_FISH:
@@ -131,10 +138,41 @@ void LOT_InitialiseSlot(int16_t item_num, int32_t slot)
         break;
 
     case O_APE:
-        creature->lot.step = STEP_L * 2;
+        creature->lot.step = WALL_L / 2;
+        creature->lot.drop = -WALL_L;
+        break;
+#else
+    case O_SHARK:
+    case O_BARRACUDA:
+    case O_DIVER:
+    case O_JELLY:
+    case O_CROW:
+    case O_EAGLE:
+        creature->lot.step = WALL_L * 20;
+        creature->lot.drop = -WALL_L * 20;
+        creature->lot.fly = STEP_L / 16;
+        if (item->object_id == O_SHARK) {
+            creature->lot.block_mask = BOX_BLOCKABLE;
+        }
+        break;
+
+    case O_WORKER_3:
+    case O_WORKER_4:
+    case O_YETI:
+        creature->lot.step = WALL_L;
         creature->lot.drop = -WALL_L;
         break;
 
+    case O_SPIDER:
+    case O_SKIDOO_ARMED:
+        creature->lot.step = WALL_L / 2;
+        creature->lot.drop = -WALL_L;
+        break;
+
+    case O_DINO:
+        creature->lot.block_mask = BOX_BLOCKABLE;
+        break;
+#endif
     default:
         break;
     }
@@ -145,9 +183,9 @@ void LOT_InitialiseSlot(int16_t item_num, int32_t slot)
     m_SlotsUsed++;
 }
 
-void LOT_CreateZone(ITEM *item)
+void LOT_CreateZone(ITEM *const item)
 {
-    CREATURE *creature = item->data;
+    CREATURE *const creature = item->data;
 
     const int16_t *zone;
     const int16_t *flip;
@@ -176,25 +214,25 @@ void LOT_CreateZone(ITEM *item)
     }
 }
 
-void LOT_InitialiseLOT(LOT_INFO *LOT)
+void LOT_InitialiseLOT(LOT_INFO *const lot)
 {
-    LOT->node =
+    lot->node =
         GameBuf_Alloc(sizeof(BOX_NODE) * Box_GetCount(), GBUF_CREATURE_LOT);
-    LOT_ClearLOT(LOT);
+    LOT_ClearLOT(lot);
 }
 
-void LOT_ClearLOT(LOT_INFO *LOT)
+void LOT_ClearLOT(LOT_INFO *const lot)
 {
-    LOT->search_num = 0;
-    LOT->head = NO_BOX;
-    LOT->tail = NO_BOX;
-    LOT->target_box = NO_BOX;
-    LOT->required_box = NO_BOX;
+    lot->search_num = 0;
+    lot->head = NO_BOX;
+    lot->tail = NO_BOX;
+    lot->target_box = NO_BOX;
+    lot->required_box = NO_BOX;
 
     for (int32_t i = 0; i < Box_GetCount(); i++) {
-        BOX_NODE *node = &LOT->node[i];
-        node->search_num = 0;
-        node->exit_box = NO_BOX;
+        BOX_NODE *const node = &lot->node[i];
         node->next_expansion = NO_BOX;
+        node->exit_box = NO_BOX;
+        node->search_num = 0;
     }
 }
