+++ src/libtrx/include/libtrx/game/pathing/lot.h
@@ -2,7 +2,11 @@
 
 #include "../creature/types.h"
 
-extern bool LOT_EnableBaddieAI(int16_t item_num, bool always);
-extern void LOT_DisableBaddieAI(int16_t item_num);
-extern CREATURE *LOT_GetBaddieSlot(int32_t i);
-extern void LOT_ClearLOT(LOT_INFO *LOT);
+void LOT_InitialiseArray(void);
+void LOT_InitialiseSlot(int16_t item_num, int32_t slot);
+void LOT_CreateZone(ITEM *item);
+void LOT_InitialiseLOT(LOT_INFO *LOT);
+bool LOT_EnableBaddieAI(int16_t item_num, bool always);
+void LOT_DisableBaddieAI(int16_t item_num);
+CREATURE *LOT_GetBaddieSlot(int32_t i);
+void LOT_ClearLOT(LOT_INFO *LOT);
